Adding signature checking
The couple tizen-platform-wrapper/tizen-platform-config is
intended to be a foundation of what is the tizen configuration.

As such, each change of the set of keys SHOULD implies a full
rebuild of tizen. It is what normally happens via the set
of build dependencies. But some developper encountered problems
because they were taking part of the system, having different
versions of configuration. Especially, it happened was installing
a fresh rpm built with GBS on an old image. This problem is reported
in the following file:

Bug-Tizen: TC-1626

This commit modify the behaviour of the generated library but
doesn't modify its API. The new principle is to add a signup that
it linked to the client and checked by the shared library. If the
client's signup misfits library's one, a critical error is printed
to using syslog and the program is aborted.

Change-Id: If234efbcf998049c51246f58e94c06f706b40983
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>