Refactoring smack labels of Homes
This evolution bring major changes to the architecture
of the home directories. The smack labels are chenged
to allow applications to run with their own label.

1. The directory TZ_USER_APP is no more labelled User.
This is not mandatory and can be revert. However, the
label User should be used with care.

2. The directory TZ_USER_DB is now labelled System::Shared
with transmuting. This allows System process to modify it
while allowing user's application to read and lock it.

3. Two new directories are set up: TZ_USER_HOME/.pki/nssdb
and TZ_USER_APP/xwalk-service. They both are needed and
used by the WRT crosswalk.

4. The desktop files are of TZ_USER_DESKTOP are managed by
the system and now are labelled System::Shared.

Change-Id: I5bd0f50e44f85395968e83d760095df45acaf8e0
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>